refactor: fixed review items
