+++ src/Xenial.Tasty/Execution/TestGroupDelegate.cs
@@ -4,5 +4,5 @@ using System.Threading.Tasks;
 
 namespace Xenial.Delicious.Execution
 {
-    public delegate Task TestDelegate(TestExecutionContext context);
+    public delegate Task TestGroupDelegate(TestGroupContext context);
 }
