+++ src/Xenial.Tasty/Execution/TestGroupMiddleware/CollectTestCaseMiddleware.cs
@@ -21,27 +21,6 @@ namespace Xenial.Delicious.Execution.TestGroupMiddleware
 
                await next();
            });
-        public static TestExecutor UseTestGroupForceVisitor(this TestExecutor executor)
-            => executor.UseGroup(async (context, next) =>
-            {
-                try
-                {
-                    if (context.CurrentGroup.IsForced != null)
-                    {
-                        var result = context.CurrentGroup.IsForced();
-
-                        foreach (var child in context.CurrentGroup.Descendants().OfType<IForceAble>())
-                        {
-                            child.IsForced = () => result;
-                        }
-                    }
-                    await next();
-                }
-                catch (Exception ex)
-                {
-                    context.CurrentGroup.Exception = ex;
-                    context.CurrentGroup.TestOutcome = TestOutcome.Failed;
-                }
-            });
+       
     }
 }
